+++ index.js
@@ -1,5 +1,6 @@
 export { default as VexAnimationController } from "./source/animation/VexAnimationController.js";
 export { default as VexEmitter } from "./source/effects/particles/VexEmitter.js";
+export { default as VexTilemap } from "./source/graphics/tile/VexTilemap.js";
 export { default as VexText } from "./source/text/VexText.js";
 export { default as VexTween } from "./source/tweens/VexTween.js";
 export { default as VexBasic } from "./source/VexBasic.js";